Update EDA.ipynb
There's a typo in the EDA.ipynb.

It would be better if you make the EDA.ipynb self explanatory with the tasks that needs to be done. In rubrics, it's mentioned to use python’s imshow to visualize medical images during EDA. 
Is it expected by students to use the test images that are inside the github repo or the files mentioned in the csv file in `/data/` folder?